trying dotnet 7 changed processor modules auto registration fixed mongo state store GetOrSet added open telemetry and mongo state store services extensions minor little things

+++ src/Krimson.State.MongoDB/MongoStateStore.cs
@@ -1,5 +1,4 @@
 ﻿using MongoDB.Driver;
-using MongoDB.Driver.Linq;
 
 namespace Krimson.State.MongoDB;
 
@@ -8,24 +7,41 @@ public class MongoStateStore : IStateStore {
         IsUpsert = true
     };
     
-    public MongoStateStore(IMongoDatabase database) => Database = database;
+    static readonly TransactionOptions DefaultTransactionOptions = new TransactionOptions().With(
+        ReadConcern.Local,
+        ReadPreference.Primary,
+        WriteConcern.W1
+    );
 
-    IMongoDatabase Database { get; }
+    static readonly ClientSessionOptions DefaultClientSessionOptions =
+        new() { DefaultTransactionOptions = DefaultTransactionOptions };
+
+    public MongoStateStore(IMongoDatabase database, string? collectionName = null) {
+        Database       = database;
+        CollectionName = collectionName ?? "krimson-state";
+    }
+
+    IMongoDatabase Database       { get; }
+    string?        CollectionName { get; }
 
     public async ValueTask Set<T>(object key, T value, CancellationToken cancellationToken = default) {
+        if (key is null) throw new ArgumentNullException(nameof(key));
+
         var state = new MongoState<T>(key.ToString()!, value, DateTimeOffset.UtcNow);
         
         await Database
-            .GetCollection<MongoState<T>>("state")
+            .GetCollection<MongoState<T>>(CollectionName)
             .ReplaceOneAsync(x => x.Key == state.Key, state, DefaultReplaceOptions, cancellationToken)
             .ConfigureAwait(false);
     }
 
     public async ValueTask<T?> Get<T>(object key, CancellationToken cancellationToken = default) {
+        if (key is null) throw new ArgumentNullException(nameof(key));
+
         var stateKey = key.ToString();
         
         return await Database
-            .GetCollection<MongoState<T>>("state")
+            .GetCollection<MongoState<T>>(CollectionName)
             .Find(x => x.Key == stateKey)
             .Project(x => x.Value)
             .FirstOrDefaultAsync(cancellationToken)
@@ -33,31 +49,35 @@ public class MongoStateStore : IStateStore {
     }
 
     public async ValueTask Delete<T>(object key, CancellationToken cancellationToken = default) {
+        if (key is null) throw new ArgumentNullException(nameof(key));
+
         var stateKey = key.ToString();
         
         await Database
-            .GetCollection<MongoState<T>>("state")
+            .GetCollection<MongoState<T>>(CollectionName)
             .DeleteOneAsync(x => x.Key == stateKey, cancellationToken: cancellationToken)
             .ConfigureAwait(false);
     }
 
     public async ValueTask<T?> GetOrSet<T>(object key, Func<ValueTask<T>> factory, CancellationToken cancellationToken = default) {
-        var stateKey = key.ToString()!;
-      
-        var exists = await Database
-            .GetCollection<MongoState<T>>("state")
-            .AsQueryable()
-            .AnyAsync(x => x.Key == stateKey, cancellationToken)
+        using var session = await Database.Client
+            .StartSessionAsync(DefaultClientSessionOptions, cancellationToken)
+            .ConfigureAwait(false);
+        
+        return await session
+            .WithTransactionAsync(Execute(), cancellationToken: cancellationToken)
             .ConfigureAwait(false);
 
-        if (exists)
-            return await Get<T>(key, cancellationToken).ConfigureAwait(false);
+        Func<IClientSessionHandle, CancellationToken, Task<T>> Execute() =>
+            async (_, ct) => {
+                var value = await Get<T>(key, ct).ConfigureAwait(false);
 
-        var value = await factory().ConfigureAwait(false);
-        var state = new MongoState<T>(stateKey, value, DateTimeOffset.UtcNow);
-
-        await Set(key, state, cancellationToken);
-
-        return value;
+                if (value is null) {
+                    value = await factory().ConfigureAwait(false);
+                    await Set(key, value, ct).ConfigureAwait(false);
+                }
+                    
+                return value;
+            };
     }
 }
